AI-ML-Internship-Eminds-feature-initial-setup/ReceipeMaker_Agent_using_LangGraph/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-groq>=0.0.1
langchain-community>=0.0.10
langchain-core>=0.1.0
python-dotenv>=1.0.0
typing-extensions>=4.0.0
pydantic>=2.0.0
